requirements.txt
# ========================================
# QUANTUM TRADING MATRIX™ - PRODUCTION REQUIREMENTS
# ========================================

# Core FastAPI and Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
starlette>=0.27.0

# Database and ORM
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Authentication and Security
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
cryptography>=41.0.7

# Redis and Caching
redis>=5.0.1
celery>=5.3.4

# Financial Data and Trading
yfinance>=0.2.24
alpha-vantage>=2.3.1
pandas>=2.1.4
numpy>=1.24.4

# Options Pricing and Risk Management
scipy>=1.11.4
scikit-learn>=1.3.2

# Data Visualization and Analysis
matplotlib>=3.8.2
seaborn>=0.13.0
plotly>=5.17.0

# Machine Learning (Optional - uncomment when ready)
# tensorflow>=2.15.0
# torch>=2.1.0
# transformers>=4.35.0

# HTTP Clients and External APIs
httpx>=0.25.2
requests>=2.31.0
aiohttp>=3.9.1

# Monitoring and Logging
prometheus-client>=0.19.0
structlog>=23.2.0
sentry-sdk[fastapi]>=1.38.0

# Development and Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1

# Environment and Configuration
python-dotenv>=1.0.0
pydantic[email]>=2.5.1
pydantic-settings>=2.1.0

# Email and Notifications
fastapi-mail>=1.4.1
jinja2>=3.1.2

# File Processing and Utilities
python-dateutil>=2.8.2
pytz>=2023.3
cachetools>=5.3.2

# Development Tools
pre-commit>=3.6.0
bandit>=1.7.5

# API Documentation Enhancement
fastapi-limiter>=0.1.6

# Additional Utilities
click>=8.1.7
python-json-logger>=2.0.7
schedule>=1.2.0

# Trading and forecasting libraries
prophet>=1.1.0
backtrader>=1.9.76.123
# pyfolio==0.9.2  # Temporarily commented due to dependency conflicts
ta>=0.10.0
statsmodels>=0.13.0
# pymc3==3.11.5  # Temporarily commented due to dependency conflicts
# skfolio==0.8.1  # Temporarily commented due to dependency conflicts

# Deep learning
torch>=2.0.0
tensorflow>=2.12.0
keras>=2.12.0
# ultralytics==8.0.145  # Temporarily commented due to dependency conflicts

# Django and REST framework
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-redis>=5.4.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
drf-yasg>=1.21.0

# Background tasks
flower>=2.0.0

# Monitoring and logging
sentry-sdk>=1.30.0
python-json-logger>=2.0.0
loguru>=0.7.0

# Testing
pytest-cov>=4.1.0
pytest-mock>=3.11.0
pytest-django>=4.7.0

# Security
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.6
django-allauth>=0.57.0

# Utilities
tqdm>=4.66.0
joblib>=1.3.0
click>=8.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pymdown-extensions>=10.0
pillow>=10.0.0 